Click history, ...

+++ lib/components/chart_data.dart
@@ -2,5 +2,5 @@ class ChartData {
   ChartData(this.duration, this.values);
 
   final String duration;
-  final double values;
+  final int values;
 }
